+++ CHANGELOG.md
@@ -4,3 +4,4 @@
     - Fixes issue with getting very few replies - https://github.com/vladkens/twscrape/commit/28ef6ba23d0e1a25064ac7c4bcd8348deb9d4a0a
     - Fetch offensive and spam replies on tweet_replies - https://github.com/vladkens/twscrape/commit/cf273ed2677322530422bb755415a607283ec7ef
     
+- Adding proxy support to xclid getter from: https://github.com/vladkens/twscrape/compare/main...Volanbober:twscrape:main